Change how the weather icon filepaths are handled
Previously, the icon filepaths were expected to come solely from
a particular folder within the project. As a result, the construct_email
function would only work if the given forecast object had an icon
file from the weather_icons folder. This has been changed to allow for
the forecast object to have an icon located anywhere on the device the
script is running from.